LSM/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0
hydra-core>=1.3.0
hydra-colorlog>=1.2.0
wandb>=0.15.0
pandas>=1.5.0
numpy>=1.24.0
Pillow>=9.5.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
boto3>=1.28.0
python-dotenv>=1.0.0
tqdm>=4.65.0
albumentations>=1.3.0
opencv-python>=4.8.0 